requirements.txt
nltk
ipython
pandas
PyYAML
scikit-learn
torch==1.8.1+cu111 torchvision==0.9.1+cu111 torchaudio==0.8.1 -f https://download.pytorch.org/whl/lts/1.8/torch_lts.html
torchtext==0.9.1
pytorch-lightning==1.3.5
tqdm
scipy
numpy
Cython
